资源分页查询bug修复

+++ src/main/Java/com/model/Page.java
@@ -12,7 +12,7 @@ public class Page<T> {
     // 每页显示记录数
     private Integer pagesize = 10;
     // 总记录数
-    private Long totalsize;
+    private Integer totalsize;
     // 数据集合
     private List<T> dataList;
 
@@ -20,7 +20,8 @@ public class Page<T> {
     }
 
     /**
-     * 用于主页面展示
+     * 用于加载主页面时展示
+     * 分页查询传入参数时使用
      * @param pageno
      */
     public Page(Integer pageno) {
@@ -29,6 +30,7 @@ public class Page<T> {
 
     /**
      * 用于用户收藏列表展示
+     * 用于获取总资源数（普通、关键字、资源类别）
      * @param pageno
      * @param pagesize
      */
@@ -37,13 +39,26 @@ public class Page<T> {
         this.pagesize = pagesize;
     }
 
-    public Page(Integer pageno, Long totalsize, List<T> dataList) {
+    /**
+     * 用于返回给前端进行主页面展示
+     * @param pageno
+     * @param totalsize
+     * @param dataList
+     */
+    public Page(Integer pageno, Integer totalsize, List<T> dataList) {
         this.pageno = pageno;
         this.totalsize = totalsize;
         this.dataList = dataList;
     }
 
-    public Page(Integer pageno, Integer pagesize, Long totalsize, List<T> dataList) {
+    /**
+     * 用于返回给前端进行用户收藏列表展示
+     * @param pageno
+     * @param pagesize
+     * @param totalsize
+     * @param dataList
+     */
+    public Page(Integer pageno, Integer pagesize, Integer totalsize, List<T> dataList) {
         this.pageno = pageno;
         this.pagesize = pagesize;
         this.totalsize = totalsize;
@@ -66,11 +81,11 @@ public class Page<T> {
         this.pagesize = pagesize;
     }
 
-    public Long getTotalsize() {
+    public Integer getTotalsize() {
         return totalsize;
     }
 
-    public void setTotalsize(Long totalsize) {
+    public void setTotalsize(Integer totalsize) {
         this.totalsize = totalsize;
     }
